pytest: add fixtures for some of the utilities in BaseBench
This purposedly duplicates most of the code from base.py as the idea is
to move all benchmarks to pytest and remove base.py entirely.

Signed-off-by: Filipe Laíns <lains@riseup.net>

pytest: fix no_scm detection in dvc fixture

Generally, we will be using this fixture with the git and git_clone
fixtures, which would have already changed the current directory to
tmpdir, but this is not guaranteed.

Signed-off-by: Filipe Laíns <lains@riseup.net>